+++ StoreClient/Entities/Zune/ZuneAlbumSearch.cs
@@ -15,15 +15,15 @@ namespace StoreClient.Entities.Zune
 
             /// <remarks/>
             [System.Xml.Serialization.XmlElementAttribute(Namespace = "http://a9.com/-/spec/opensearch/1.1/")]
-            public byte startIndex { get; set; }
+            public int startIndex { get; set; }
 
             /// <remarks/>
             [System.Xml.Serialization.XmlElementAttribute(Namespace = "http://a9.com/-/spec/opensearch/1.1/")]
-            public byte totalResults { get; set; }
+            public int totalResults { get; set; }
 
             /// <remarks/>
             [System.Xml.Serialization.XmlElementAttribute(Namespace = "http://a9.com/-/spec/opensearch/1.1/")]
-            public byte itemsPerPage { get; set; }
+            public int itemsPerPage { get; set; }
 
             /// <remarks/>
             public DateTime updated { get; set; }
@@ -103,7 +103,7 @@ namespace StoreClient.Entities.Zune
 
             /// <remarks/>
             [System.Xml.Serialization.XmlElementAttribute(Namespace = "http://schemas.zune.net/catalog/music/2007/10")]
-            public byte playRank { get; set; }
+            public int playRank { get; set; }
 
             /// <remarks/>
             [System.Xml.Serialization.XmlElementAttribute(Namespace = "http://schemas.zune.net/catalog/music/2007/10")]
